pyproject.toml
[project]
name = "awsurance-product-calculation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "sympy>=1.13.3",
]
